renamed test project
