+++ app/src/test/java/at/aau/wagnis/ObjectTest/PlayerTest.java
@@ -1,4 +1,4 @@
-package at.aau.wagnis.AdditionalTest;
+package at.aau.wagnis.ObjectTest;
 
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 import static org.junit.jupiter.api.Assertions.assertEquals;
